fix header loading of bg color, cleanup some var names - converting member vars to be prefixed with underscore

+++ ninja-engine/window.h
@@ -99,7 +99,7 @@ class GameWindow {
 										int bottom_y, int top_y, 
 										int level_height );
 
-		void SetClearColor(uint r, uint g, uint b);
+		void SetClearColor(float r, float g, float b);
 
 		void BeginDrawing();
 		void EndDrawing();
